algorithm-deps-install (#1081)
* add dependencyInstallCmd option to algorithm
add python support

* add nodejs support

* fix install deps script permission

* add test

+++ core/algorithm-builder/environments/nodejs/dockerfile/requirements.sh
@@ -2,11 +2,11 @@
 
 set -e
 
-PACKAGES_REGISTRY=${packagesRegistry}
+export PACKAGES_REGISTRY=${packagesRegistry}
 PACKAGES_REGISTRY_HOST=${PACKAGES_REGISTRY#http://}
-PACKAGES_REGISTRY_HOST=${PACKAGES_REGISTRY_HOST#https://}
-PACKAGES_TOKEN=${packagesToken}
-PACKAGES_AUTH=${packagesAuth}
+export PACKAGES_REGISTRY_HOST=${PACKAGES_REGISTRY_HOST#https://}
+export PACKAGES_TOKEN=${packagesToken}
+export PACKAGES_AUTH=${packagesAuth}
 
 if [ ! -z ${PACKAGES_REGISTRY} ]; then
     echo "found npm registry ${PACKAGES_REGISTRY}"
@@ -16,6 +16,14 @@ if [ ! -z ${PACKAGES_REGISTRY} ]; then
         echo "//${PACKAGES_REGISTRY_HOST}:_auth=${PACKAGES_AUTH}" > ${HOME}/.npmrc
         echo "//${PACKAGES_REGISTRY_HOST}:always-auth=true" >> ${HOME}/.npmrc
     fi
+fi
+if [ ! -z ${dependency_install_cmd} ]; then
+    echo "found dependency install script"
+    SCRIPT_CWD="${PWD}"
+    echo "running ${dependency_install_cmd} in folder ${SCRIPT_CWD}"
+    sh -c "cd ${SCRIPT_CWD} && sh ${dependency_install_cmd}"
+    echo "${dependency_install_cmd} execution done with code $?"
+elif [ ! -z ${PACKAGES_REGISTRY} ]; then
     npm install --registry=${PACKAGES_REGISTRY}
     rm -f .npmrc
 else
